refactor: 컨트롤러 패키지 구조 변경
