+++ src/main/java/roomescape/controller/reservation/AdminReservationController.java
@@ -1,4 +1,4 @@
-package roomescape.controller.admin;
+package roomescape.controller.reservation;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
