+++ src/main/java/roomescape/controller/reservation/ReservationController.java
@@ -1,4 +1,4 @@
-package roomescape.controller;
+package roomescape.controller.reservation;
 
 import java.util.List;
 import java.util.Optional;
